bot/requirements.txt
aiogram>=3.4.1
aiohttp>=3.9.3
pydantic>=2.4.1,<2.10.0
pydantic-settings>=2.2.1
sqlalchemy>=2.0.28
asyncpg>=0.29.0
python-dotenv>=1.0.1
google-genai
pillow
